requirements.txt
Flask>=3.0.0
Flask-SocketIO>=5.3.0
docker>=7.0.0
mysql-connector-python>=8.3.0
paramiko>=3.4.0
pyserial>=3.5
requests>=2.31.0
Werkzeug>=3.0.0
esptool>=4.7.0
eventlet>=0.35.0